Create This script generates a list of random passwords.
head command:  print the top N number of data of the given input. By default, it prints the first 10 lines of the specified files. 
fold command: fold command in Linux wraps each line in an input file to fit a specified width and prints it to the standard output. By    default, it wraps lines at a maximum width of 80 columns but this is configurable.  -w :By using this option in fold command, we can limit the width by number of columns.

shuf command: The shuf command generates random permutations from input lines to standard output.