Create This script generates a random password for each user specified on the comand line.
basename:  basename strips directory information and suffixes from file names i.e. it prints the file name NAME with any leading directory components removed.

dirname:  dirname is a command in Linux which is used to remove the trailing forward slahes “/” from the NAME and prints the remaining portion. If the argument NAME does not contains the forward slash “/” then it simply prints dot “.”.

${0}: returns path of the recently used directory.