Create Demonstrate the use of shift and while loops.
shift is a builtin command of the Bash shell. When executed, it shifts the positional parameters (such as arguments passed to a bash script) to the left, putting each parameter in a lower position.
When you run shift, the current positional parameters are shifted left n times. Positional parameter x is given the value of parameter x+n. If parameter x+n does not exist, parameter x is unset.
**For better understanding run the script.**
If not specified, the default value of n is 1. So the commands "shift 1" and "shift" (with no argument) do the same thing.
If a parameter is shifted to a position with a number less than 1, it "falls off" — its value is discarded. So the command shift always discards the previous value of $1, and shift 2 always discards the previous values of $1 and $2.
The special positional parameter $0 is excluded from all shift operations, and never modified by the shift command.